+++ CHANGELOG.md
@@ -4,6 +4,7 @@ FEATURES:
 
 * **New Resource:** `gigahost_dns_record`
 * **New Resource:** `gigahost_dns_zone`
+* **New Resource:** `gigahost_ssh_key`
 * **New Data Source:** `gigahost_account`
 * **New Data Source:** `gigahost_dns_records`
 * **New Data Source:** `gigahost_dns_zones`